Move and rename simpleAnimation to animation
